NPC Rat EQS & Chase Prototype
BT_NPC_Rat now has defined functionalities for Notice Mode and Passive Mode. When a player is seen, the rat enters Notice Mode, using an EQS system to determine which spot is farthest away from the player and scurrying towards it. When a player is not seen for 3 seconds, the rat enters Passive Mode and stands still. These are controlled by the rat's blackboard and an enum determining its state.