BPI Capturable & Jenny Pounce - 2nd Pass
A new interface called BPI_Capturable has been created. This interface is used for any capturable pawns, such as rats, but it may also be expanded to work with kittens and other collectibles in the future. When Jenny pounces, she creates a sphere trace for objects at her feet. Any rat detected in this radius is turned off.