Mini-Map Widget - 1st Pass
Logic for a mini-map widget has been created, but it is turned off by default. By pressing Right Click or clicking the Right Stick Button while the the mini-map widget is hidden, it is made visible, and vice versa. The icons are placeholders and are intended to be replaced by icons made by the art team.