Reinstated Old Capture Deletion on Rats
The old method for turning off Rat actors now works with the mini-map, and the rat is not captured multiple times when turned off. With these changes, this rat removal method has been reinstated rather than deleting the whole actor. It is currently unknown which method between this and actor deletion is more performant, but it is currently believed to be the former due to less garbage collection and ease of turnoff in case of reloads.