Wall Cling - 1st Pass
A wall cling function for movement is in the works. Movement functions have been slightly altered to work with this system. By pressing Left Ctrl or Right Trigger, Jenny will check to see if there is a wall in front of her. If there is a wall, she will be able to move along it until the line trace returns negative. There is currently a bug where Jenny gets stuck in Falling Mode, which will be resolved shortly.