BPI_Dismount
A new BPI_Jenny function called BPI_Dismount has been created. By pouncing on a wall while attached to it, Jenny will dismount the wall and fall back down. This ensures players will not be stuck on a wall and be forced to climb it, but this system is still tied to the pounce cooldown. Further testing to see if this rule should stay may be needed.