Objective Display - 1st Pass
When entering the all new Objective Trigger actor blueprint, an Objective Display pop-up will appear in the top left corner of the screen. This data is represented by a struct that contains two text variables for the name and description, two int variables for the current value of the objective and the value needed to clear the objective, and a boolean that controls whether or not the objective has been cleared. This is passed through two functions for adding an objective to the player and setting the objective the player must clear if it does not already exist in the struct array.