Collision/Mesh Fixes & Experiments With Player Camera
Jenny's model has been adjusted to keep her on the floor instead of slightly hovering above it. The NPC Rat colliders have been adjusted in height, and their proxy limbs no longer collide with the player after their capture. The length of Jenny's camera boom/spring arm have been reduced from 400 to 250, bringing the camera closer to Jenny and providing an improved sense of scale, but adjustments are still needed when it comes to panning the camera towards the floor and bouncing off walls.